+++ sample/src/main/java/com/synerise/sdk/sample/model/LoyaltyPointsContent.java
@@ -16,5 +16,4 @@ public class LoyaltyPointsContent {
     public void setPoints(String points) {
         this.points = points;
     }
-
 }