+++ sample/src/main/java/com/synerise/sdk/sample/ui/section/adapter/SectionsPagerAdapter.java
@@ -3,7 +3,6 @@ package com.synerise.sdk.sample.ui.section.adapter;
 import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentManager;
 
-import com.synerise.sdk.injector.ui.walkthrough.pager.InfinitePagerAdapter;
 import com.synerise.sdk.sample.data.Section;
 
 import java.util.List;