+++ sample/src/main/java/com/synerise/sdk/sample/ui/settings/SettingsFragment.java
@@ -77,7 +77,5 @@ public class SettingsFragment extends BaseFragment {
         } else {
             signOut.setVisibility(GONE);
         }
-
-        view.findViewById(R.id.show_walkthrough).setOnClickListener(v -> Injector.showWalkthrough());
     }
 }